Fix data handling in System Exclusive event
The event specifies its data length with an initial varlen.  Currently
this length is incorrectly copied into the output, and all input bytes
up to the next F7 are consumed, and (except fo the F7) copied into the
output.  This is incorrect: F7 is not considered to terminate the
message unless it occurs at the position stated by the length.