Handle F7 Sysex events
A System Exclusive event may start with F7 instead of F0.  These F7
messages sometimes occur as continuations of the previous F0 message, if
the F0 message did not end with an F7 byte.
They can also be used as an "escape", to bundle arbitrary data into the
input.

For the current application, it suffices to handle F7 messages the same
way as F0 messages.