Fix example of writing CSV string to file
As you mention [here](https://github.com/timwedde/py_midicsv/issues/10#issuecomment-770933885), the `midi_to_csv` function returns a list of strings. `write` is expecting a string rather than a list of strings, so running the example results in an error. Your provided fix was to use the `writelines` function instead, but you forgot to update the README.